requirements.txt
attrs==21.4.0
cftime==1.5.1.1
click==8.0.3
importlib-metadata==4.8.3
netCDF4==1.5.8
numpy==1.19.5
pandas==1.1.5
pdoc==0.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
scipy==1.5.4
six==1.16.0
typing_extensions==4.0.1
zipp==3.6.0
